Fix the turn as int

+++ student.py
@@ -196,7 +196,7 @@ class Piggy(pigo.Pigo):
                 largest_dist = distance
                 ang = index
         print("I think the best angle is %d\n" % ang)
-        turn = 7 * abs(ang - self.MIDPOINT) / 90   ##calculate how much it should turn to the valid direction.
+        turn = int(7 * abs(ang - self.MIDPOINT) / 90)   ##calculate how much it should turn to the valid direction.
         if ang <= self.MIDPOINT:
             self. encR(turn)
         if ang > self.MIDPOINT:
